Finished mandelbrot set.

+++ fractals-java/src/Colour.java
@@ -3,11 +3,11 @@ public class Colour {
 	public static final Colour WHITE = new Colour(1,1,1);
 	public static final Colour RED   = new Colour(1,0,0);
 
-	private final float r;
-	private final float g;
-	private final float b;
+	private final double r;
+	private final double g;
+	private final double b;
 
-	public Colour(float r, float g, float b) {
+	public Colour(double r, double g, double b) {
 		if ((r < 0.0f) || (r > 1.0f)) throw new IllegalArgumentException("r is out of range");
 		if ((g < 0.0f) || (g > 1.0f)) throw new IllegalArgumentException("g is out of range");
 		if ((b < 0.0f) || (b > 1.0f)) throw new IllegalArgumentException("b is out of range");
@@ -18,19 +18,19 @@ public class Colour {
 	}
 	
 	// Constructor for grayscale colors
-	public Colour(float c) {
+	public Colour(double c) {
 		this(c,c,c);
 	}
 
-	public float getR() {
+	public double getR() {
 		return r;
 	}
 
-	public float getG() {
+	public double getG() {
 		return g;
 	}
 	
-	public float getB() {
+	public double getB() {
 		return b;
 	}
 }
